requirements.txt
python-telegram-bot==20.3
Pillow==10.1.0